restructure handlers so code DRY is followed
